moved all utility functions to utility.js and changed the imports.
