+++ cypress/e2e/wip/Tradershub/accountSignupDIEL(MF+CR).cy.js
@@ -1,5 +1,5 @@
 import '@testing-library/cypress/add-commands'
-import { generateEpoch } from '../../../support/tradersHub'
+import { generateEpoch } from '../../../support/helper/utility'
 
 const regulationText = '.regulators-switcher__switch div.item.is-selected'
