+++ cypress/e2e/wip/Tradershub/addDemoFinancialExistingMT5Pwd.cy.js
@@ -1,5 +1,5 @@
 import '@testing-library/cypress/add-commands'
-import { generateEpoch } from '../../../support/tradersHub'
+import { generateEpoch } from '../../../support/helper/utility'
 
 describe('QATEST-5724: CFDs - Create a demo Financial account using existing MT5 account password', () => {
   const signUpEmail = `sanity${generateEpoch()}mail@deriv.com`
