+++ cypress/e2e/wip/Tradershub/mt5FinancialDemoSignup.cy.js
@@ -1,5 +1,5 @@
 import '@testing-library/cypress/add-commands'
-import { generateEpoch } from '../../../support/tradersHub'
+import { generateEpoch } from '../../../support/helper/utility'
 
 describe('QATEST-5699: Create a Financial Demo CFD account', () => {
   const signUpEmail = `sanity${generateEpoch()}mt5financialdemo@deriv.com`
