+++ cypress/support/commands/index.js
@@ -1,9 +1,6 @@
 import './tradersHub'
 import './dBot'
 import './dTrader'
-import './cashier'
 import './common'
-import './reports'
-import './settings'
 import './wallets'
 import './p2p'
